# be/cmd/server/docs/swagger.yaml
...
info:
  contact: {}
paths:
  /api/assets:
    get:
      consumes:
      - application/json
      description: Get all assets
      produces:
      - application/json
      responses: {}
      summary: Get all assets
      tags:
      - Assets
    post:
      consumes:
      - multipart/form-data
      description: Create assets
      parameters:
      - description: Asset Name
        in: formData
        name: assetName
        required: true
        type: string
      - description: Purchase Date (RFC3339 format, e.g. 2023-04-15T10:00:00Z)
        in: formData
        name: purchaseDate
        required: true
        type: string
      - description: Cost
        in: formData
        name: cost
        required: true
        type: number
      - description: Warranty Expiry (RFC3339 format, e.g. 2023-12-31T23:59:59Z)
        in: formData
        name: warrantExpiry
        required: true
        type: string
      - description: Serial Number
        in: formData
        name: serialNumber
        required: true
        type: string
      - description: Category ID
        in: formData
        name: categoryId
        required: true
        type: integer
      - description: Department ID
        in: formData
        name: departmentId
        required: true
        type: integer
      - description: redirect url
        in: formData
        name: redirectUrl
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Image to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses: {}
      summary: Create assets
      tags:
      - Assets
...
  /api/assets-retired/{id}:
    patch:
      consumes:
      - application/json
      description: Retired assets
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Retired assets
      tags:
      - Assets
  /api/assets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete assets
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete assets
      tags:
      - Assets
    get:
      consumes:
      - application/json
      description: Get assets
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get assets
      tags:
      - Assets
    put:
      consumes:
      - multipart/form-data
      description: Update assets
      parameters:
      - description: Asset Name
        in: formData
        name: assetName
        required: true
        type: string
      - description: Purchase Date (RFC3339 format, e.g. 2023-04-15T10:00:00Z)
        in: formData
        name: purchaseDate
        required: true
        type: string
      - description: Cost
        in: formData
        name: cost
        required: true
        type: number
      - description: Warranty Expiry (RFC3339 format, e.g. 2023-12-31T23:59:59Z)
        in: formData
        name: warrantExpiry
        required: true
        type: string
      - description: maintenance_time Float e.g. 7.5
        in: formData
        name: maintenance_time
        required: true
        type: number
      - description: Serial Number
        in: formData
        name: serialNumber
        required: true
        type: string
      - description: Serial Number
        in: formData
        name: status
        required: true
        type: string
      - description: Category ID
        in: formData
        name: categoryId
        required: true
        type: integer
      - description: Department ID
        in: formData
        name: departmentId
        required: true
        type: integer
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Image to upload
        in: formData
        name: image
        required: true
        type: file
      - description: expectDayMaintenance Date (RFC3339 format, e.g. 2023-04-15T10:00:00Z)
        in: formData
        name: expectDayMaintenance
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Update assets
      tags:
      - Assets
  /api/assets/filter:
    get:
      consumes:
      - application/json
      description: Get all assets have permission
      parameters:
      - in: query
        name: assetName
        type: string
      - in: query
        name: categoryId
        type: string
      - in: query
        name: cost
        type: string
      - in: query
        name: departmentId
        type: string
      - in: query
        name: email
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: serialNumber
        type: string
      - in: query
        name: status
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: Get all assets with filter
      tags:
      - Assets
  /api/assets/filter-dashboard:
    get:
      consumes:
      - application/json
      description: Get Dashboard
      parameters:
      - in: query
        name: categoryId
        type: string
      - in: query
        name: departmentId
        type: string
      - description: '"csv" hoặc "pdf" hoặc ""'
        in: query
        name: export
        type: string
      - in: query
        name: status
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: Get Dashboard
      tags:
      - Assets
...
swagger: "2.0"